+++ client/src/hooks/post/like/usePostLikes.ts
@@ -1,6 +1,6 @@
 import { FEEDNEST_BACKEND_API } from "@/utils/apiClient";
 import { useMutation } from "@tanstack/react-query";
-import type { IGenericMessageResponse } from "../../../types/dist/index";
+import type { IGenericMessageResponse } from "../../../../../types/dist";
 
 const handlePostLikes = async (postId:string): Promise<IGenericMessageResponse> => {
   try {
