+++ client/src/hooks/user/useUserDataUpdate.ts
@@ -1,6 +1,6 @@
 import { FEEDNEST_BACKEND_API } from "@/utils/apiClient";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import type { IUserDataResponse } from "../../../types/dist/index";
+import type { IUserDataResponse } from "../../../../types/dist";
 import type { TUserDataUpdate } from "@/utils/schema/userData";
 
 const handleUserDataUpdate = async (
